Sensor v2.1.0

+++ Sensor/src/homography/homography.h
@@ -17,8 +17,8 @@
             homography();
             void calculateHomographyMatrix();
             void calculateCoordinates(int, int);
-            int getX();
-            int getY();
+            uint16_t getX();
+            uint16_t getY();
             void setCalArray(int,int,int);
             int getCalArray(int, int);
             void orderCalArray();
@@ -26,8 +26,8 @@
         private:
             uint8_t findMax(uint16_t ARRAY[]);
             mtx_type _H[3][3];
-            int _x;
-            int _y;
+            uint16_t _x;
+            uint16_t _y;
             uint16_t _bounds[2][4];
             uint16_t _calArray[2][4];
     };
